# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/catalog:
    get:
      operationId: getTokenCatalog
      tags: [token-purchases]
      summary: Get the GLD and SLV catalog
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Catalog with current approved delayed quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenCatalogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /wallet/inventory:
    get:
      operationId: getTokenInventory
      tags: [token-purchases]
      summary: Get escrow-backed token inventory
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Inventory and approved quote availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenInventoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /wallet/purchase:
    post:
      operationId: purchaseToken
      tags: [token-purchases]
      summary: Spend 100 gems for one GLD or SLV token payout
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CsrfToken"
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenPurchaseInput"
      responses:
        "200":
          description: Existing or newly-created purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPurchaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Idempotency key already belongs to another symbol, or no token inventory
        "503":
          description: Approved quote or payout infrastructure unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/encounter-ticket:
    post:
      operationId: issueEncounterTicket
      tags: [token-rewards]
      summary: Issue a single-use combat encounter eligibility ticket
      security: [{ walletSession: [] }]
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EncounterTicketInput" }
      responses:
        "201":
          description: Ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EncounterTicketResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Another encounter ticket is active }
  /wallet/encounter-ticket/{id}/monster:
    post:
      operationId: settleEncounterMonster
      tags: [token-rewards]
      summary: Settle one defeated monster against an encounter ticket
      security: [{ walletSession: [] }]
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CsrfToken"
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EncounterMonsterInput" }
      responses:
        "200": { description: Settlement result }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /wallet/rewards:
    get:
      operationId: getTokenRewards
      tags: [token-rewards]
      summary: Get accumulated server-side monster rewards
      security: [{ walletSession: [] }]
      parameters: [{ $ref: "#/components/parameters/WalletAddress" }]
      responses:
        "200":
          description: Accumulated GLD and SLV rewards
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenRewardsResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /wallet/rewards/claim:
    post:
      operationId: claimTokenRewards
      tags: [token-rewards]
      summary: Claim all unclaimed monster rewards
      security: [{ walletSession: [] }]
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Durable reward claim status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenRewardClaimResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /wallet/redeemable-gems:
    get:
      operationId: getRedeemableGemBalance
      tags: [token-purchases]
      summary: Get the server-authoritative token-redeemable gem balance
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Redeemable gem balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemableGemBalanceResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /wallet/purchase/{id}:
    get:
      operationId: getTokenPurchase
      tags: [token-purchases]
      summary: Get a purchase status
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
        - in: path
          name: id
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: Purchase status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPurchaseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Purchase not found
  /wallet/purchases/pending:
    get:
      operationId: getPendingTokenPurchases
      tags: [token-purchases]
      summary: Get pending purchases for the authenticated wallet
      security:
        - walletSession: []
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Pending purchase statuses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPurchasePendingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    walletSession:
      type: apiKey
      in: cookie
      name: dicebound_session
  parameters:
    WalletAddress:
      in: header
      name: X-Wallet-Address
      required: true
      schema:
        type: string
        pattern: "^0x[0-9a-fA-F]{40}$"
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Wallet session is missing or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Origin, wallet, or CSRF check failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    TokenSymbol:
      type: string
      enum: [GLD, SLV]
    TokenCatalogItem:
      type: object
      required: [symbol, tokenAddress, decimals, gemCost, usdValue, tokenAmount, quotePrice, quoteSource, quoteTimestamp, quoteDelayed, delayed, available]
      properties:
        symbol: { $ref: "#/components/schemas/TokenSymbol" }
        tokenAddress: { type: string }
        decimals: { type: integer, const: 18 }
        gemCost: { type: integer, const: 100 }
        usdValue: { type: string, const: "10" }
        tokenAmount: { type: string }
        quotePrice: { type: string }
        quoteSource: { type: string }
        quoteTimestamp: { type: string, format: date-time }
        quoteDelayed: { type: string, const: "previous-close" }
        delayed: { type: boolean, const: true }
        available: { type: boolean }
    TokenCatalogResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TokenCatalogItem" }
    TokenInventoryItem:
      allOf:
        - $ref: "#/components/schemas/TokenCatalogItem"
        - type: object
          required: [escrowBalance, escrowBalanceBaseUnits, reservedBaseUnits, availableBalanceBaseUnits]
          properties:
            escrowBalance: { type: string }
            escrowBalanceBaseUnits: { type: string }
            reservedBaseUnits: { type: string }
            availableBalanceBaseUnits: { type: string }
    TokenInventoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TokenInventoryItem" }
    TokenPurchaseInput:
      type: object
      required: [symbol]
      properties:
        symbol: { $ref: "#/components/schemas/TokenSymbol" }
    EncounterTicketInput:
      type: object
      required: [encounterId, monsterIds]
      properties:
        encounterId: { type: string, minLength: 1, maxLength: 120 }
        monsterIds:
          type: array
          minItems: 1
          maxItems: 12
          items: { type: string }
    EncounterTicketResponse:
      type: object
      required: [id, encounterId, monsterIds, expiresAt]
      properties:
        id: { type: string }
        encounterId: { type: string }
        monsterIds: { type: array, items: { type: string } }
        expiresAt: { type: string, format: date-time }
    EncounterMonsterInput:
      type: object
      required: [monsterId]
      properties:
        monsterId: { type: string }
    TokenRewardsResponse:
      type: object
      required: [enabled, GLD, SLV, GLDValueCents, SLVValueCents, count]
      properties:
        enabled: { type: boolean }
        GLD: { type: string }
        SLV: { type: string }
        GLDValueCents: { type: integer, minimum: 0 }
        SLVValueCents: { type: integer, minimum: 0 }
        count: { type: integer, minimum: 0 }
    TokenRewardClaimResponse:
      type: object
      required: [id, status, GLD, SLV, pending]
      properties:
        id: { type: ["string", "null"] }
        status: { type: string }
        GLD: { type: string }
        SLV: { type: string }
        pending: { type: boolean }
    TokenPurchaseResponse:
      type: object
      required: [id, symbol, status, gemCost, tokenAmount, quotePrice, quoteSource, quoteTimestamp, quoteDelayed, delayed, redeemableGemBalance]
      properties:
        id: { type: string }
        symbol: { $ref: "#/components/schemas/TokenSymbol" }
        status:
          type: string
          enum: [pending, signed, submitted, confirmed, failed]
        gemCost: { type: integer, const: 100 }
        tokenAmount: { type: string }
        quotePrice: { type: string }
        quoteSource: { type: string }
        quoteTimestamp: { type: string, format: date-time }
        quoteDelayed: { type: string, const: "previous-close" }
        delayed: { type: boolean, const: true }
        redeemableGemBalance: { type: integer, minimum: 0 }
        transactionHash: { type: ["string", "null"] }
    RedeemableGemBalanceResponse:
      type: object
      required: [balance]
      properties:
        balance: { type: integer, minimum: 0 }
    TokenPurchasePendingResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TokenPurchaseResponse" }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }